em-resources.ts
