+++ src/camera/camera.ts
@@ -1,5 +1,6 @@
 import { CanvasDef } from "../render/canvas.js";
-import { Component, EM, EntityW, Resource } from "../ecs/entity-manager.js";
+import { Component, EM, EntityW } from "../ecs/entity-manager.js";
+import { Resource } from "../ecs/em-resources.js";
 import { V3, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { MeDef } from "../net/components.js";
 import { WorldFrameDef } from "../physics/nonintersection.js";
