+++ src/ecs/em-init.ts
@@ -6,15 +6,11 @@ import {
 import { resetTempMatrixBuffer } from "../matrix/sprig-matrix.js";
 import { getCallStack, assert } from "../utils/util-no-import.js";
 import { isPromise } from "../utils/util.js";
-import {
-  ResourceDef,
-  ComponentDef,
-  Resources,
-  componentsToString,
-  ResId,
-  _em,
-  _resources,
-} from "./entity-manager.js";
+import { _resources } from "./em-resources.js";
+import { ComponentDef, componentsToString, _em } from "./entity-manager.js";
+import { Resources } from "./em-resources.js";
+import { ResId } from "./em-resources.js";
+import { ResourceDef } from "./em-resources.js";
 
 export type InitFnId = number;
