+++ src/graybox/graybox-ship-arena.ts
@@ -4,7 +4,8 @@ import { AlphaDef, ColorDef, TintsDef } from "../color/color-ecs.js";
 import { ENDESGA16, seqEndesga16 } from "../color/palettes.js";
 import { DevConsoleDef } from "../debug/console.js";
 import { DeadDef, DeletedDef } from "../ecs/delete.js";
-import { EM, Entity, EntityW, Resources } from "../ecs/entity-manager.js";
+import { EM, Entity, EntityW } from "../ecs/entity-manager.js";
+import { Resources } from "../ecs/em-resources.js";
 import { createEntityPool } from "../ecs/entity-pool.js";
 import { Phase } from "../ecs/sys-phase.js";
 import { createHexGrid, hexXYZ, hexesWithin } from "../hex/hex.js";
