+++ src/hyperspace/game-hyperspace.ts
@@ -1,5 +1,6 @@
 import { CameraDef } from "../camera/camera.js";
-import { EM, EntityW, Resources } from "../ecs/entity-manager.js";
+import { EM, EntityW } from "../ecs/entity-manager.js";
+import { Resources } from "../ecs/em-resources.js";
 import { PositionDef, RotationDef, ScaleDef } from "../physics/transform.js";
 import { RendererDef, RenderableConstructDef } from "../render/renderer-ecs.js";
 import { blurPipelines } from "../render/pipelines/std-blur.js";
