+++ src/input/inputs.ts
@@ -1,5 +1,6 @@
 import { Canvas, CanvasDef } from "../render/canvas.js";
-import { Component, EM, Resource } from "../ecs/entity-manager.js";
+import { Component, EM } from "../ecs/entity-manager.js";
+import { Resource } from "../ecs/em-resources.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { clamp } from "../utils/math.js";
 import { DEBUG_INPUTS } from "../flags.js";
