+++ src/levels/level-map.ts
@@ -1,4 +1,5 @@
-import { EM, Resource } from "../ecs/entity-manager.js";
+import { EM } from "../ecs/entity-manager.js";
+import { Resource } from "../ecs/em-resources.js";
 import { VERBOSE_LOG } from "../flags.js";
 import { TextDef } from "../gui/ui.js";
 import { AABB2, aabbCenter2, updateAABBWithPoint2_ } from "../physics/aabb.js";
