+++ src/net/components.ts
@@ -1,4 +1,5 @@
-import { EM, Component, Resource } from "../ecs/entity-manager.js";
+import { EM, Component } from "../ecs/entity-manager.js";
+import { Resource } from "../ecs/em-resources.js";
 import { Deserializer } from "../utils/serialize.js";
 import { MessageType } from "./message.js";
 import { FromNetworkEvent, ToNetworkEvent } from "./network-events.js";
