+++ src/physics/phys-helpers.ts
@@ -1,10 +1,6 @@
-import {
-  ComponentDef,
-  EntityW,
-  EM,
-  ResourceDef,
-  Resources,
-} from "../ecs/entity-manager.js";
+import { ComponentDef, EntityW, EM } from "../ecs/entity-manager.js";
+import { Resources } from "../ecs/em-resources.js";
+import { ResourceDef } from "../ecs/em-resources.js";
 import { Phase } from "../ecs/sys-phase.js";
 import { PhysicsResultsDef } from "./nonintersection.js";
