+++ src/wind/wind.ts
@@ -1,4 +1,5 @@
-import { Component, EM, Resource } from "../ecs/entity-manager.js";
+import { Component, EM } from "../ecs/entity-manager.js";
+import { Resource } from "../ecs/em-resources.js";
 import { randInt } from "../utils/math.js";
 import { RendererDef } from "../render/renderer-ecs.js";
 import { V, V3 } from "../matrix/sprig-matrix.js";
